Combine conftest.py.
