+++ test/conftest.py
@@ -8,6 +8,7 @@ from threading import enumerate as thread_enumerate
 import pytest
 
 from pymodbus.datastore import ModbusBaseSlaveContext
+from pymodbus.transport import NULLMODEM_HOST, CommParams, CommType
 from pymodbus.transport.transport import NullModem
 
 
@@ -40,6 +41,58 @@ BASE_PORTS = {
 def get_base_ports():
     """Return base_ports."""
     return BASE_PORTS
+
+
+@pytest.fixture(name="use_comm_type")
+def prepare_dummy_use_comm_type():
+    """Return default comm_type."""
+    return CommType.TCP
+
+
+@pytest.fixture(name="use_host")
+def define_use_host():
+    """Set default host."""
+    return NULLMODEM_HOST
+
+
+@pytest.fixture(name="use_cls")
+def prepare_commparams_server(use_port, use_host, use_comm_type):
+    """Prepare CommParamsClass object."""
+    if use_host == NULLMODEM_HOST and use_comm_type == CommType.SERIAL:
+        use_host = f"{NULLMODEM_HOST}:{use_port}"
+    return CommParams(
+        comm_name="test comm",
+        comm_type=use_comm_type,
+        reconnect_delay=0,
+        reconnect_delay_max=0,
+        timeout_connect=0,
+        source_address=(use_host, use_port),
+        baudrate=9600,
+        bytesize=8,
+        parity="E",
+        stopbits=2,
+    )
+
+
+@pytest.fixture(name="use_clc")
+def prepare_commparams_client(use_port, use_host, use_comm_type):
+    """Prepare CommParamsClass object."""
+    if use_host == NULLMODEM_HOST and use_comm_type == CommType.SERIAL:
+        use_host = f"{NULLMODEM_HOST}:{use_port}"
+    timeout = 10 if not pytest.IS_WINDOWS else 2
+    return CommParams(
+        comm_name="test comm",
+        comm_type=use_comm_type,
+        reconnect_delay=0.1,
+        reconnect_delay_max=0.35,
+        timeout_connect=timeout,
+        host=use_host,
+        port=use_port,
+        baudrate=9600,
+        bytesize=8,
+        parity="E",
+        stopbits=2,
+    )
 
 
 @pytest.fixture(name="system_health_check", autouse=True)
@@ -193,14 +246,3 @@ class mockSocket:  # pylint: disable=invalid-name
     def setblocking(self, _flag):
         """Set blocking."""
         return None
-
-
-_CURRENT_PORT = 5200
-
-
-@pytest.fixture(name="use_port")
-def get_port():
-    """Get next port."""
-    global _CURRENT_PORT  # pylint: disable=global-statement
-    _CURRENT_PORT += 1
-    return _CURRENT_PORT
